--- components/index.tsx
@@ -1,1 +1,0 @@
-export { default as Brand } from './Brand/Brand';
