+++ package.json
@@ -11,7 +11,7 @@
   "dependencies": {
     "@react-navigation/native": "^6.1.6",
     "@react-navigation/native-stack": "^6.9.12",
-    "@types/react": "~18.0.27",
+    "@types/react": "^18.2.0",
     "expo": "~48.0.10",
     "expo-splash-screen": "~0.18.1",
     "expo-status-bar": "~1.4.4",
@@ -23,10 +23,14 @@
     "react-native-web": "~0.18.10",
     "react-navigation": "^4.4.4",
     "readme-md-generator": "^1.0.0",
-    "typescript": "^4.9.4"
+    "typescript": "^5.0.4"
   },
   "devDependencies": {
-    "@babel/core": "^7.20.0"
+    "@babel/core": "^7.20.0",
+    "@tsconfig/react-native": "^3.0.0",
+    "@types/jest": "^29.5.1",
+    "@types/react-test-renderer": "^18.0.0",
+    "babel-plugin-module-resolver": "^5.0.0"
   },
   "private": true
 }
