+++ screens/Details.tsx
@@ -1,0 +1,14 @@
+import * as React from 'react';
+import { Button, View, Text } from "react-native";
+
+export function DetailsScreen({ navigation }) {
+  return (
+    <View style={{ flex: 1, alignItems: "center", justifyContent: "center" }}>
+      <Text>Details Screen</Text>
+      <Button
+        title="Go to Details... again"
+        onPress={() => navigation.navigate("Details")}
+      />
+    </View>
+  );
+}
