+++ screens/Home.tsx
@@ -1,0 +1,38 @@
+import * as React from "react";
+import { StyleSheet, TouchableOpacity, View, Text } from "react-native";
+
+import Brand from "../components/Brand/Brand";
+
+export function HomeScreen({ navigation }) {
+  return (
+    <View style={{ flex: 1, alignItems: "center", justifyContent: "center" }}>
+      <Brand />
+      <TouchableOpacity
+        style={styles.button}
+        onPress={() => navigation.navigate("Details")}
+      >
+        <Text style={styles.text}>Access</Text>
+      </TouchableOpacity>
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  button: {
+    alignItems: "center",
+    justifyContent: "center",
+    paddingVertical: 12,
+    paddingHorizontal: 32,
+    borderRadius: 8,
+    marginTop: 10,
+    elevation: 3,
+    backgroundColor: "black",
+  },
+  text: {
+    fontSize: 16,
+    lineHeight: 21,
+    fontWeight: "bold",
+    letterSpacing: 0.25,
+    color: "white",
+  },
+});
