+++ tsconfig.json
@@ -1,4 +1,16 @@
 {
   "compilerOptions": {},
-  "extends": "expo/tsconfig.base"
+  "extends": "expo/tsconfig.base",
+  "baseUrl": ".",
+  "paths": {
+    "*": [
+      "src/*"
+    ],
+    "tests": [
+      "tests/*"
+    ],
+    "@components/*": [
+      "src/components/*"
+    ],
+  },
 }
